Move "Deathly Hallow Race" content into a separate view
The "Deathly Hallow Race" section in the Home/Index.cshtml has been moved to its own view called Home/DeathlyHallow.cshtml. This has been done to decouple the content and to make the pages more manageable. Some alterations were made to notifications.js to modify the structure of displayed message and count for improved readability. A "Deathly Hallows" link is also added to the navigation bar.