Integrate Azure SignalR Service to application
The changes in the code reflect the configuration needed for leveraging the Azure SignalR Service in our SignalRExamples application. A connection string for Azure SignalR has been added in the appsettings.json (*removed sensitive data*) file and also in the services configuration. The Azure SignalR package has been added to the SignalRExamples.csproj.